+++ electron/preload.ts
@@ -68,6 +68,8 @@ contextBridge.exposeInMainWorld('electronAPI', {
   openInIDE: (ideCmd: string, dirPath: string) => ipcRenderer.invoke('shell:openInIDE', ideCmd, dirPath),
   detectClaude: () => ipcRenderer.invoke('shell:detectClaude'),
   openClaude: (dirPath: string, skipPerms: boolean) => ipcRenderer.invoke('shell:openClaude', dirPath, skipPerms),
+  detectCodex: () => ipcRenderer.invoke('shell:detectCodex'),
+  openCodex: (dirPath: string, fullAuto: boolean) => ipcRenderer.invoke('shell:openCodex', dirPath, fullAuto),
   openPath: (path: string) => ipcRenderer.invoke('shell:openPath', path),
   openInObsidian: (path: string) => ipcRenderer.invoke('shell:openInObsidian', path),
   startDrag: (filePath: string) => ipcRenderer.send('native:startDrag', filePath),